+++ src/App.vue
@@ -1,11 +1,11 @@
 <template>
   <div id="app">
-    <Screener :screener="screener" />
+    <VueScreener :screener="screener" />
   </div>
 </template>
 
 <script setup lang="ts">
-import Screener from './components/Screener.vue'
+import VueScreener from './components/VueScreener.vue'
 import { useVueScreener } from './hooks/use-vue-screener'
 
 const screener = useVueScreener([])
