+++ src/index.ts
@@ -1,4 +1,4 @@
-import Screener from './components/Screener.vue'
+import VueScreener from './components/VueScreener.vue'
 import ScreenerPagination from './components/ScreenerPagination.vue'
 import ScreenerSearch from './components/ScreenerSearch.vue'
 import ScreenerSearchOptions from './components/ScreenerSearchOptions.vue'
@@ -6,4 +6,4 @@ import TableCell from './components/ui/table/TableCell.vue'
 import { useVueScreener } from './hooks/use-vue-screener'
 import './assets/tailwind.css'
 
-export { Screener, ScreenerSearch, ScreenerSearchOptions, ScreenerPagination, TableCell, useVueScreener }
+export { VueScreener, ScreenerSearch, ScreenerSearchOptions, ScreenerPagination, TableCell, useVueScreener }
