+++ src/stories/1-basic-usage/1-array-of-objects.story.vue
@@ -1,13 +1,13 @@
 <template>
   <Story title="1. Array of objects" source="-">
     <DevSandbox title="Array of objects" :screener="screener">
-      <Screener :screener="screener" />
+      <VueScreener :screener="screener" />
     </DevSandbox>
   </Story>
 </template>
 
 <script lang="ts" setup>
-import { Screener, useVueScreener } from '../../index'
+import { VueScreener, useVueScreener } from '../../index'
 import DevSandbox from '../../components/dev/Sandbox.vue'
 import baseData from '../../fixtures/data.json'
