+++ src/stories/1-basic-usage/2-array-of-primitives.story.vue
@@ -1,13 +1,13 @@
 <template>
   <Story title="2. Array of primitives" source="-">
     <DevSandbox title="Array of primitives" :screener="screener">
-      <Screener :screener="screener" />
+      <VueScreener :screener="screener" />
     </DevSandbox>
   </Story>
 </template>
 
 <script lang="ts" setup>
-import { Screener, useVueScreener } from '../../index'
+import { VueScreener, useVueScreener } from '../../index'
 import DevSandbox from '../../components/dev/Sandbox.vue'
 import primitivesData from '../../fixtures/primitives-data.json'
