+++ src/stories/1-basic-usage/3-array-of-mixed-objects.story.vue
@@ -1,13 +1,13 @@
 <template>
   <Story title="3. Array of mixed objects" source="-">
     <DevSandbox title="Array of mixed objects" :screener="screener">
-      <Screener :screener="screener" />
+      <VueScreener :screener="screener" />
     </DevSandbox>
   </Story>
 </template>
 
 <script lang="ts" setup>
-import { Screener, useVueScreener } from '../../index'
+import { VueScreener, useVueScreener } from '../../index'
 import DevSandbox from '../../components/dev/Sandbox.vue'
 import mixedObjectsData from '../../fixtures/mix-objects-data.json'
