+++ src/stories/1-basic-usage/4-loading.story.vue
@@ -1,13 +1,13 @@
 <template>
   <Story title="4. Loading" source="-">
     <DevSandbox title="Loading" :screener="screener">
-      <Screener :screener="screener" />
+      <VueScreener :screener="screener" />
     </DevSandbox>
   </Story>
 </template>
 
 <script lang="ts" setup>
-import { Screener, useVueScreener } from '../../index'
+import { VueScreener, useVueScreener } from '../../index'
 import DevSandbox from '../../components/dev/Sandbox.vue'
 import baseData from '../../fixtures/data.json'
