+++ src/stories/1-basic-usage/6-bad-data.story.vue
@@ -1,13 +1,13 @@
 <template>
   <Story title="6. Bad data" source="-">
     <DevSandbox title="Bad data" :screener="screener">
-      <Screener :screener="screener" />
+      <VueScreener :screener="screener" />
     </DevSandbox>
   </Story>
 </template>
 
 <script lang="ts" setup>
-import { Screener, useVueScreener } from '../../index'
+import { VueScreener, useVueScreener } from '../../index'
 import DevSandbox from '../../components/dev/Sandbox.vue'
 
 const badData: any = 'bad data'
