+++ src/stories/1-basic-usage/7-sort-default-sort-directions.story.vue
@@ -1,7 +1,7 @@
 <template>
   <Story title="7. Sorting / Default sort directions" source="-">
     <DevSandbox title="Sorting / Default sort directions" :screener="screener">
-      <Screener
+      <VueScreener
         :screener="screener"
         :ui="{
           tableView: {
@@ -18,7 +18,7 @@
 </template>
 
 <script lang="ts" setup>
-import { Screener, useVueScreener } from '../../index'
+import { VueScreener, useVueScreener } from '../../index'
 import DevSandbox from '../../components/dev/Sandbox.vue'
 import baseData from '../../fixtures/data.json'
