+++ src/stories/4-other/1-stress-test.story.vue
@@ -1,13 +1,13 @@
 <template>
   <Story title="1. Stress Test" source="-">
     <DevSandbox title="Results" :screener="screener">
-      <Screener :screener="screener" />
+      <VueScreener :screener="screener" />
     </DevSandbox>
   </Story>
 </template>
 
 <script lang="ts" setup>
-import { Screener, useVueScreener } from '../../index'
+import { VueScreener, useVueScreener } from '../../index'
 import baseData from '../../fixtures/stress-data.json'
 import DevSandbox from '../../components/dev/Sandbox.vue'
